Massive overhaul to project:
App.java now loads menu.fxml first.

Sudoku.java was changed from an interface to an abstract class that extends Observable. All functions changed to abstract functions.

SudokuImpl.java was updated to have a startGame() function that now calls the createGrid() function (initializes game).

Update.java was changed to GameState.java, and the enums within were changed to RUNNING, NOT_RUNNING, COMPLETED, CHECK, and CHOICE. These keep track of the state of the game so the controller/view can be updated properly.

SudokuSingleton.java was changed to be a Sudoku and it returns an instance of a SudokuImpl.

Util.java was created that has a setRoot() function and a loadFXML() function. These are specifically for the App.java class as well as the View classes.

GameView.java now loads the game.fxml. This is used in MenuController.java when the startGame() function is called by pressing the "Start Game" button on the main menu. It creates an instance of Sudoku through the singleton and GameView. We add the gameView as an observer of the instance and set the gameview scene into the stage and start the game.