Made extreme overhaul to the entire project.
App.java is now simplified to just load the menu utilizing the scene size constants from the View.Util class, and launch the program into the loaded menu.

Difficulty.java was created as an enum that holds 3 constant variables for an easy, medium, and hard difficulty. Later on these will be utilized to change the amount of numbers that are removed from the solution to create the puzzle.

GameController.java is now properly separated for MVC structure, and contains an instance of the Sudoku class (no longer SudokuSingleton, now Sudoku.getInstance()) and the GameView. The controller also handles the pausing of the game, as well as contains a timer using the Timer class and the TimerTask class, which are imported from java.util. When the game starts, the timer will schedule the timerTask at a fixed rate, which will be in seconds. The timerTask will call run() and will update the view and model every second, and check for an event using the eventHandler with the gameView. The only current event is the escape key, however this will contain the select option for selecting a number from the hot bar on the bottom of the gameView, and then selecting a box in the playSpace to place the number.

GameOverPane.java is a StackPane that is placed inside the GameView whenever the game is over. The Pane is set to contain an opaque background, and display a title that states "Game Over", as well as showing the score and time it took to complete the game. There are two buttons, one to play again, and another to go to the main menu.

GameState.java had a few changes. There is now a PAUSED state, which is stating that the PausePane is up and the timer is paused.

GameView is now created to contain a canvas with GraphicsContext. We create a playSpace (GridPane, this is where the puzzle is that you can place numbers into), and a numberSelectionArea, or hot bar (this is where the numbers are that you can select to place within the playSpace). There is two GridPanes called scoreBox and timeBox, that will be stacked on top of each other to display the score and the time. These GridPane's will contain two labels each. One label is the text for the word score and time (scoreLabel and timeLabel), and the other is to display the actual score and time (scoreField and timeField). The hud (BorderPane) then adds these children. Currently, the GameView does not properly display, however it will in the next commit.

game.fxml has been removed and will be created programmatically using GameView.java.

menu.fxml has been revamped to be more organized and also now display a ComboBox for choosing a difficulty.

MenuController.java now has an initialize function that sets the game difficulty based on what was selected, and default selects "Easy". The startGame function is called when the menu button "Start Game" is pressed (onAction). It sets the difficulty in the sudoku.getInstance() and will create the gameView and gameController as well as set the scene on the stage containing the gameView and starts the game through the gameController. There are also functions created for the quit button, howtoplay button, settings button, and leaderboard button.

PausePane.java is similar to GameOverPane.java where it has an opaque background and displays that the game is Paused as well as contains a button to exit the game and return to the menu. This will end the game, not save the score to the user (needs to be created) and will return to the menu.

Sudoku.java had a few minor bugs and didn't contain all the needed functions for the model to properly start. These bugs have been fixed. I changed the way that GameState works, and we're no longer using Observable and Observer the way that we were before, but instead we're using the GameState as a way to observe and update. Score is now set to start at 1000 and will eventually be modified to decrease by 10 point for every 10 seconds that the game takes, and 10 points will be subtracted for every mistake made in the puzzle (compared to the solution). Instead of a singleton class, I have created a getInstance() function that returns an instance of the model. There is also a setDifficulty function that utilizes the enums.

Util.java now just contains some constants for the width/height of the gameView as well as for font and the getStage() function which utilizes an action event. This is for switching to the main menu from the game.