Major work done to the Firebase database.
Created FirebaseService.java which establishes connection to the firebase database. Utilizes the provided secret token to connect to the database and get an instance of the database.

Created ShowDbChanges.java which displays any changes to the database in the console when it occurs. Later on, once data is able to be written, or users are able to be created and login/logout, whenever something happens to the database, this will display it in the console.

Created FirebaseAuthentication.java, which will be used for creating accounts, logging in, and logging out of the accounts.