Finished all fxml files. Changed the gameview to be an fxml that uses GameController as its controller.
Created controllers for all finished FXML files that properly switch between the views when buttons and specific actions are completed.

Heavily modified DatabaseController to work perfectly.

Modified App.java to open the login screen instead of the menu.

Modified Sudoku.java to work properly with the way I'd like the game to run.

Being very bland with commit comments as I'm in a rush at the moment but the next commit will be more detailed on what changes occurred here and with that push.