Final game changes. Edited the DatabaseController to now have score be a value that the user will add to, rather than a high score.
Changed GameController to have a Timeline that updates every second. Each second it will call updateLabels(). The timeline also allows for game pausing and ending the game. updateLabels does all the view updating, such as changing the value of the cell on the board to the selected number that is placed at that location. It will also handle communicating with the model to make the changes that it makes to the view.

Updated GameState to match the changes made to the model. The only states we have now are Running, NotRunning, Paused, and Completed. NotRunning is default, Running is when the game has started and during gameplay, Paused is whenever escape is pressed and the timeline is paused (can't play the game), and Completed is whenever the game is finished (puzzle complete) or when the game is over after failing all attempts.

GameOverPane is now a window that pops up whenever the game ends. If you win the game, it says "You won!" and gives your score and how long it took to complete the game. If you lost due to failing all attempts, it says "Game Over" and your score never changes.